delete from remote post

+++ journal/models/comment.py
@@ -34,6 +34,9 @@ class Comment(Content):
 
     @classmethod
     def update_by_ap_object(cls, owner, item, obj, post_id, visibility):
+        p = cls.objects.filter(owner=owner, item=item).first()
+        if p and p.edited_time >= datetime.fromisoformat(obj["updated"]):
+            return p  # incoming ap object is older than what we have, no update needed
         content = obj.get("content", "").strip() if obj else ""
         if not content:
             cls.objects.filter(owner=owner, item=item).delete()
